+++ src/main/java/fr/tmsconsult/p3_backend_chatop/services/impl/JwtServiceImpl.java
@@ -5,6 +5,7 @@ import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
+import jakarta.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
@@ -85,5 +86,13 @@ public class JwtServiceImpl implements IJwtService {
     public Date extractExpiration(String token) {
         return extractClaim(token, Claims::getExpiration);
     }
+    public String extractTokenFromRequest(HttpServletRequest request) {
+        String token = "";
+        String authHeader = request.getHeader("Authorization");
+        if (authHeader != null && authHeader.startsWith("Bearer ")) {
 
+            token = authHeader.substring(7);
+        }
+        return token;
+    }
 }
